requirements.txt
altair==6.0.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.2
cairocffi==1.7.1
CairoSVG==2.8.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
cssselect2==0.8.0
defusedxml==0.7.1
docx2pdf==0.1.8
fire==0.7.1
fonttools==4.61.0
gitdb==4.0.12
GitPython==3.1.45
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lxml==6.0.2
MarkupSafe==3.0.3
narwhals==2.13.0
numpy==2.2.6
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.3
pdf2docx==0.5.8
pdfminer.six==20251107
pillow==12.0.0
pluggy==1.6.0
protobuf==6.33.1
pyarrow==22.0.0
pycparser==2.23
pydeck==0.9.1
Pygments==2.19.2
PyMuPDF==1.26.6
PyPDF2==3.0.1
pytest==9.0.1
pytest-mock==3.15.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
pywin32==311
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
smmap==5.0.2
streamlit==1.52.0
tenacity==9.1.2
termcolor==3.2.0
tinycss2==1.5.1
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
webencodings==0.5.1
